requirements_vector.txt
# LangChain Vector Matching Dependencies
# Using free/cheap embedding options first

# Core LangChain
langchain>=0.1.0
langchain-community>=0.0.20
langchain-google-genai>=1.0.0  # For Gemini (free tier)
langchain-openai>=0.0.5  # For OpenAI (fallback, more expensive)

# Vector Stores (free)
chromadb>=0.4.0  # Vector database (free, local)
faiss-cpu>=1.7.4  # Fast vector search (free, local)

# Embedding Models (free/cheap options)
google-generativeai>=0.3.0  # Gemini embeddings (free tier)
sentence-transformers>=2.2.0  # Free local embeddings (no API needed!)

# Data Processing
pandas>=2.0.0
duckdb>=0.9.0
pyyaml>=6.0

# Utilities
python-dotenv>=1.0.0
rich>=13.0.0

# Optional: For Firework/GLM if needed
# requests>=2.31.0  # For custom API calls

